Configuration: added some missing proxy settings options.
There are some http related settings that are set internally but not
exposed to the config system.

Add

  - proxy_buffer_size
  - proxy_header_buffer_size
  - proxy_timeout
  - proxy_send_timeout
  - proxy_read_timeout

to the http configuration in nxt_router.c and the http configuration
validation in nxt_conf_validation.c

This doesn't add 'proxy_buffers' as that is unused and will be removed
in a follow up commit.

This then allows users to set these such as

   "settings": {
       "http": {
           "proxy_buffer_size": 65536,
           "proxy_header_buffer_size": 8192,
           "proxy_timeout": 30,
           "proxy_send_timeout": 60,
           "proxy_read_timeout": 60
       }
   }

If these values aren't set they retain their default values.