Var: translate $$ to a literal $.
Allow $$ to translate to a literal $ to allow support for sub-delimiters
in URIs.

It is possible to have URLs like

  https://example.com/path/15$1588/9925$2976.html

and thus it would be useful to be able to specify them in various bits
of the unit config such as the location setting.

However this hadn't been possible due to $ being used to denote
variables for substitution. E.g $host.

As was suggested by @alejandro-colomar in the below GitHub issue, a nice
way to handle this would be to use $$ to represent a literal $.

This means the above URL could be specified as

  https://example.com/path/15$$1588/9925$$2976.html

in the unit config.

This is done by adding a variable called '$' which is loaded into the
variables hash table which translates into a literal $.

This case is then handled in nxt_var_next_part() where variables are
parsed for lookup and $$ is set for substitution by a single $. Actual
variable substitution happens in nxt_var_query_finish().

This will work anywhere where variable substitution works and of course
you can mix normal variables with these such as

  https://$host/path/$$$uri/index.html

which might translate to

  https://example.com/path/$/old/uri/index.html

Also, due to $$ now being a valid construct, we need to remove some no
longer invalid tests (we do it here to avoid breaking the tests between
commits).

Closes: https://github.com/nginx/unit/issues/675