Build: added a make check target.
While unit comes with a test suite, it is not entirely obvious how to
run it!.

As suggested by @alejandro-colomar add a make check target which is a
common target for running such things. This is added to the Makefile in
the repository root. It will build unit first if required.

On FreeBSD and Fedora, pytest can be run as pytest or pytest-3, on
Debian it is pytest-3, so configure tries to find the right command to
run. If it is called other names on other systems they can be added to
the PYTEST_CMDS array.