Added const to string pointers that never change.
Constant string pointers that don't change the string to which
they point can use const to help the compiler create optimal code.

This patch doesn't affect at all the current resulting binary,
because since our strings are stored in static variables, the
compiler ha enough information to know that it can make them
const, but it's still a good thing to add it, because the compiler
may not always be able to optimize things, and also readers can
see that the pointers never point to other things.

===

See the sizes of the binary before and after this patch
(`size -B build/unitd > ...`):

$ cat sz.B.ptr
   text	   data	    bss	    dec	    hex	filename
 436246	  29648	   1224	 467118	  720ae	build/unitd
$ cat sz.B.const
   text	   data	    bss	    dec	    hex	filename
 436246	  29648	   1224	 467118	  720ae	build/unitd