Storing string literals as 'char []' instead of 'char *'.
If a string is stored as an array, the binary only contains the
string literal.  But if it is stored as a pointer, the binary
needs to contain the string literal _and_ a pointer to it, which
means an extra 8 bytes for every string.

In the case of arrays of strings, the size of the inner arrays
(the strings) has to be the same for all of them, so we need, at
least, the size required by the longest string (strlen + 1).

===

See the sizes of the binary before and after this patch
(`size -B build/unitd > ...`):

$ cat sz.B.const
   text	   data	    bss	    dec	    hex	filename
 436246	  29648	   1224	 467118	  720ae	build/unitd
$ cat sz.B.arr
   text	   data	    bss	    dec	    hex	filename
 434846	  29008	   1224	 465078	  718b6	build/unitd

Both 'text' and 'data' have decreased after this patch.  Good.
text by 1.4 KB (0.3%), and data by 0.6 KB (2%).

===

See a more detailed difference of the sizes
(`size -A build/unitd | sed -E 's/^(\S+\s+\S+)\S+\s+/\1/' > ...`):

$ diff -u sz.A.{const,arr}
--- sz.A.const	2022-06-12 20:32:15.029643211 +0200
+++ sz.A.arr	2022-06-12 20:33:12.187521769 +0200
@@ -9,20 +9,20 @@
 .dynstr                 7151
 .gnu.version             932
 .gnu.version_r           352
-.rela.dyn              27984
+.rela.dyn              26328
 .rela.plt               3744
 .init                     23
 .plt                    2512
 .plt.got                   8
 .text                 264353
 .fini                      9
-.rodata                57626
+.rodata                57882
 .eh_frame_hdr           8220
 .eh_frame              49488
 .tbss                    560
 .init_array                8
 .fini_array                8
-.data.rel.ro           14192
+.data.rel.ro           13552
 .dynamic                 560
 .got                      88
 .got.plt                1272
@@ -30,13 +30,13 @@
 .bss                     664
 .comment                  30
 .debug_aranges          4656
-.debug_info          1610066
+.debug_info          1610114
 .debug_abbrev         112065
-.debug_line           250577
+.debug_line           250579
 .debug_str             67547
 .debug_line_str         4061
 .debug_loclists       282985
 .debug_rnglists        12075
-Total                2811180
+Total                2809190

Ignoring debug sections, we see that:

.rela.dyn    decreased by ~ 1.6 KB (6%)
.rodata      INCREASED by ~ 0.3 KB (0.6%)
.data.rel.ro decreased by ~ 0.6 KB (4%)

This also looks great.  Most of the changes are for good, and
there's a section that increases, but not very much (and .rodata
will decrease again in the last patch of this patch set,
"Merged identical strings into unique extern constant variables.").