Aligned arrays of strings to a power of 2.
Arrays of strings are being stored as arrays of arrays of char.
Since arrays can't have padding, the compiler is forced to create
arrays of arrays of the size we specify.  If that size is not a
power of 2, the compiler is able to align the outer array to the
next power of two, but it can't help with the alignment of the
inner arrays.

To improve performance when getting pointers to a string contained
in an array of strings, we can add some padding ourselves, by
rounding up the array size to a power of two.

Since the strings are not huge, the padding shouldn't cause too
much wasted space, and it's still far from the waste that an array
of pointers would cause.

Although this should have an impact in the binary size, it
doesn't, which makes me think this code is being unused.

===

See the sizes of the binary before and after this patch
(`size -B build/unitd > ...`):

$ cat sz.B.arr
   text	   data	    bss	    dec	    hex	filename
 434846	  29008	   1224	 465078	  718b6	build/unitd
$ cat sz.B.pow
   text	   data	    bss	    dec	    hex	filename
 434846	  29008	   1224	 465078	  718b6	build/unitd

===

Although our code doesn't show an improvement, see the following
experiment, which demonstrates the improvement:

```c
$ cat array_of_strings7.c
extern const char s[][7];

const char *foo(int i)
{
	return s[i];
}
```
```asm
$ cat array_of_strings7.s
	.file	"array_of_strings.c"
	.text
	.p2align 4
	.globl	foo
	.type	foo, @function
foo:
.LFB0:
	.cfi_startproc
	movslq	%edi, %rdi
	leaq	s(%rip), %rdx
	leaq	0(,%rdi,8), %rax
	subq	%rdi, %rax
	addq	%rdx, %rax
	ret
	.cfi_endproc
.LFE0:
	.size	foo, .-foo
	.ident	"GCC: (Debian 10.2.1-6) 10.2.1 20210110"
	.section	.note.GNU-stack,"",@progbits

```
```c
$ cat array_of_strings8.c
extern const char s[][8];

const char *foo(int i)
{
	return s[i];
}
```
```asm
$ cat array_of_strings8.s
	.file	"array_of_strings8.c"
	.text
	.p2align 4
	.globl	foo
	.type	foo, @function
foo:
.LFB0:
	.cfi_startproc
	movslq	%edi, %rdi
	leaq	s(%rip), %rax
	leaq	(%rax,%rdi,8), %rax
	ret
	.cfi_endproc
.LFE0:
	.size	foo, .-foo
	.ident	"GCC: (Debian 10.2.1-6) 10.2.1 20210110"
	.section	.note.GNU-stack,"",@progbits
```