Merged identical strings into unique extern constant variables.
Although the linker is smart enough to merge most of the
duplicated code, it's better to don't duplicate source code
(unless it has significant advantages).

In this case, we can see that we get a small improvement after
this patch.

===

See the sizes of the binary before and after this patch
(`size -B build/unitd > ...`):

$ cat sz.B.pow
   text	   data	    bss	    dec	    hex	filename
 434846	  29008	   1224	 465078	  718b6	build/unitd
$ cat sz.B.ext
   text	   data	    bss	    dec	    hex	filename
 434590	  29008	   1224	 464822	  717b6	build/unitd

'text' has decrased by 0.35 KB (0.08%).

===

See a more detailed difference of the sizes
(`size -A build/unitd | sed -E 's/^(\S+\s+\S+)\S+\s+/\1/' > ...`):

$ diff -u sz.A.{pow,ext}
--- sz.A.pow	2022-06-12 21:06:19.925747567 +0200
+++ sz.A.ext	2022-06-12 21:06:19.925747567 +0200
@@ -16,7 +16,7 @@
 .plt.got                   8
 .text                 264353
 .fini                      9
-.rodata                57882
+.rodata                57626
 .eh_frame_hdr           8220
 .eh_frame              49488
 .tbss                    560
@@ -29,12 +29,12 @@
 .data                  13520
 .bss                     664
 .comment                  30
-.debug_aranges          4656
-.debug_info          1610114
-.debug_abbrev         112065
-.debug_line           250579
-.debug_str             67547
-.debug_line_str         4061
-.debug_loclists       282985
+.debug_aranges          4688
+.debug_info          1611151
+.debug_abbrev         112343
+.debug_line           250702
+.debug_str             67555
+.debug_line_str         4095
+.debug_loclists       282939
 .debug_rnglists        12075
-Total                2809190
+Total                2810400

Ignoring debug sections, we see that:

.rodata decreased by 0.35 KB (6%).

Which is the exact amount that it increased in the previous patch
"Storing string literals as 'char []' instead of 'char *'.".

It's good that after the complete patch set transforming strings
no section (other than the debug ones) has seen a net increase.