Moved nxt_calendar variables into a structure.
This reduces the binary size, probably because the compiler can
derivate the address of one of the variables, from the address of
the other, with an offset from the same structure.

===

See the sizes of the binary before and after this patch
(`size -B build/unitd > ...`):

$ cat sz.B.ext
   text	   data	    bss	    dec	    hex	filename
 434590	  29008	   1224	 464822	  717b6	build/unitd
$ cat sz.B.struct
   text	   data	    bss	    dec	    hex	filename
 434574	  29008	   1224	 464806	  717a6	build/unitd

'text' has decreased by 16 B (0.004%).

This seems to confirm the prediction that the compiler derivates
one address from the other, since nxt_calendar.wday is used in
exactly two places (8 * 2 = 16).

===

See a more detailed difference of the sizes
(`size -A build/unitd | sed -E 's/^(\S+\s+\S+)\S+\s+/\1/' > ...`):

$ diff -u sz.A.{ext,struct}
--- sz.A.ext	2022-06-12 21:06:19.925747567 +0200
+++ sz.A.struct	2022-06-12 23:15:12.114790906 +0200
@@ -14,7 +14,7 @@
 .init                     23
 .plt                    2512
 .plt.got                   8
-.text                 264353
+.text                 264337
 .fini                      9
 .rodata                57626
 .eh_frame_hdr           8220
@@ -30,11 +30,11 @@
 .bss                     664
 .comment                  30
 .debug_aranges          4688
-.debug_info          1611151
-.debug_abbrev         112343
-.debug_line           250702
-.debug_str             67555
+.debug_info          1611450
+.debug_abbrev         112329
+.debug_line           250700
+.debug_str             67570
 .debug_line_str         4095
-.debug_loclists       282939
+.debug_loclists       282961
 .debug_rnglists        12075
-Total                2810400
+Total                2810704

Ignoring debug sections, we see that:

.text decreased by 16 B (0.006%).

===

I didn't move mday <src/nxt_time_parse.c> into the structure too,
because it increases the binary size, which looks like it removes
some optimization opportunity (which makes sense, because we're
changing a variable from being static to being extern).