Removed --no-unix-socket flag.
That flag has never worked.  The code wrapped by the corresponding
preprocessor macro NXT_HAVE_UNIX_DOMAIN has never compiled with
clang(1), due to an accident in a refactor from 2017.  That proves
that nodoby really uses the flag, since we haven't had any report
about Unit not compiling with this flag enabled, until I tried to
compile it one day and found the serious problem.

Now that we fixed the flag to at least compile, we still have to
consider if we want it.  Specifically, I don't trust some code
that hasn't been even compiled (not saying about running it or
running the whole test suite) in 5 years.

I don't trust this code at all, and no users seems to require it,
so I think it makes sense to just remove the option.  It will
simplify existing code, and also will make it easier to write new
code (or edit the existing one), and code that is easy to read and
write usually has less bugs.